Remove mock data file `mock_data.go` from tests directory to clean up unused code and reduce repository size.
